[ADD] #52466 - Download attached file in hd5
Now there is a possibility to add a virtual directory to Helpdesk 5, if theres a virtual directory on the customer, the links of files will go to this file.
If theres no one, it will check in tblglobalsettings if theres one there, it will use it.

Else it will use physical path.

Upload will still be using the physical path.